Tasker was deleted as it was usless, list of tasks were moved to Account class
